+++ packages/pro/src/ai/index.ts
@@ -4,5 +4,6 @@ export { parseResponseFormat } from "./models/openai"
 export { getLLM, getLLMConfig, getLLMOrThrow, LLMRequest } from "./llm"
 export { LLM } from "./models/base"
 export * from "./prompts"
+export * from "./reasoningFallback"
 export * from "./generators"
 export * from "./structuredOutputs"
